refactor: snsLoginView로 분리 (#91)
