+++ DietLog-MVVM/View/SignIn/SignInViewController.swift
@@ -11,10 +11,10 @@ import RxSwift
 
 class SignInViewController: BaseViewController {
     
+    // MARK: - UI Component
     private lazy var dietLogLabel = UILabel()
     private lazy var signInView = SignInView()
-    private lazy var stackVew = UIStackView()
-    private lazy var snsTitleLabel = UILabel()
+    private lazy var snsLoginView = SnsLoginView()
 
     // MARK: - 변수
     private let viewModel = SignInViewModel()
@@ -34,48 +34,19 @@ class SignInViewController: BaseViewController {
     
     // MARK: - Setup UI
     override func setupUI() {
-        view.addSubviews([dietLogLabel, signInView, snsTitleLabel, stackVew])
+        view.addSubviews([dietLogLabel, signInView, snsLoginView])
+        
+        setupDietLogLabelUI()
         signInView.configure()
-        setupLabelslUI()
-        setupStackView()
+        snsLoginView.configure()
     }
     
-    private func setupLabelslUI() {
+    private func setupDietLogLabelUI() {
         dietLogLabel.configure(text: SignInText.dietLog, font: .largeTitle)
         dietLogLabel.textAlignment = .center
         dietLogLabel.textColor = .customGreen
-        
-        snsTitleLabel.configure(text: SignInText.sns, font: .boldBody)
-        snsTitleLabel.textAlignment = .center
     }
 
-    private func setupStackView() {
-        stackVew.axis = .vertical
-        stackVew.spacing = 0
-        stackVew.alignment = .center
-        stackVew.distribution = .equalCentering
-        
-        let kakaoButton = UIButton()
-        kakaoButton.setTitle("카카오로 가입하기", for: .normal)
-        kakaoButton.addTarget(self, action: #selector(loginWithKakao), for: .touchUpInside)
-        
-        let naverButton = UIButton()
-        naverButton.setTitle("네이버로 가입하기", for: .normal)
-        naverButton.addTarget(self, action: #selector(loginWithNaver), for: .touchUpInside)
-        
-        [kakaoButton, naverButton].forEach {
-            stackVew.addArrangedSubview($0)
-            
-            $0.setTitleColor(.black, for: .normal)
-            $0.titleLabel?.font = .smallBody
-            
-            $0.snp.makeConstraints { make in
-                make.height.equalTo(40)
-                make.width.equalToSuperview()
-            }
-        }
-    }
-    
     // MARK: - Setup Layout
     override func setupLayout() {
         
@@ -85,24 +56,14 @@ class SignInViewController: BaseViewController {
         }
         
         signInView.snp.makeConstraints { make in
-            make.centerY.equalToSuperview()
+            make.top.equalTo(dietLogLabel.snp.bottom).offset(48)
             make.leading.trailing.equalToSuperview().inset(Padding.leftRightSpacing.rawValue)
             make.height.equalTo(view.snp.height).dividedBy(2)
         }
         
-        dietLogLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.bottom.equalTo(signInView.snp.top).offset(-48)
-        }
-
-        snsTitleLabel.snp.makeConstraints { make in
+        snsLoginView.snp.makeConstraints { make in
+            make.leading.trailing.equalTo(signInView)
             make.top.equalTo(signInView.snp.bottom).offset(24)
-            make.bottom.equalTo(stackVew.snp.top).offset(-8)
-            make.leading.trailing.equalTo(24)
-        }
-        
-        stackVew.snp.makeConstraints { make in
-            make.leading.trailing.equalTo(snsTitleLabel)
         }
     }
